dedupe some types
